Refactor and code layout cleaning
