+++ app/src/main/res/layout/activity_splash.xml
@@ -8,7 +8,7 @@
         tools:context=".presentation.splash.SplashActivity">
 
     <com.airbnb.lottie.LottieAnimationView
-            android:id="@+id/splashAnimation"
+            android:id="@+id/splashAnimationView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginStart="30dp"
